Move components into Dashboard file
