+++ src/components/Dashboard/Dashboard.css
@@ -1,0 +1,53 @@
+body {
+  background-image: url('ron-kenney-vG3iYginN30-unsplash.jpg');
+  background-size: cover;
+  background-repeat: no-repeat;
+  color: white;
+  margin: 0;
+  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
+    'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
+    sans-serif;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+}
+
+a {
+  text-decoration: none;
+  color: white;
+}
+
+button {
+  border: none;
+  height: 8%;
+  width: 20%;
+  background-color: #3c6658;
+  color: white;
+  font-weight: bold;
+  box-shadow: 3px 3px 3px rgba(0, 0, 0, 0.25);
+}
+
+button:hover {
+  cursor: pointer;
+  background-color: #7aa396;
+  border-radius: 3px;
+}
+
+.appt-block {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  background-color: #FF9F1C;
+  width: 70vw;
+  height: 65vh;
+  margin-top: 2%;
+  padding: 2%;
+}
+
+.main-appts {
+  width: 100%;
+  height: 70%;
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+  grid-row-gap: 2%;
+  grid-column-gap: 2%;
+}
