+++ src/components/NavBar/NavBar.css
@@ -1,9 +1,8 @@
 .nav-bar {
   display: flex;
-  width: 100%;
+  width: 100vw;
   height: 15vh;
   background-color:rgb(40, 156, 146, 0.8);
-
   margin: 0;
   padding: 0;
 }
